code/legal-llm/requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-neo4j>=0.1.0
langchain-experimental>=0.0.47
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-google-genai>=0.0.6

# PDF Processing (better structure preservation)
pdfplumber>=0.10.3
PyMuPDF>=1.23.0
pypdf>=3.17.0

# NLP and Entity Recognition
spacy>=3.7.0
spacy-transformers>=1.3.0

# Neo4j
neo4j>=5.14.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
tqdm>=4.66.0

# Optional but recommended
beautifulsoup4>=4.12.0
lxml>=4.9.0
